Generate multiple files for multiple packages
