+++ go.mod
@@ -7,6 +7,5 @@ require (
 	github.com/google/go-cmp v0.3.0
 	github.com/jinzhu/inflection v1.0.0
 	github.com/lfittl/pg_query_go v1.0.0
-	github.com/lib/pq v1.2.0 // indirect
 	github.com/spf13/cobra v0.0.5
 )
